+++ src/Rib.hs
@@ -2,6 +2,7 @@
 module Rib
   ( module Rib.App,
     module Rib.Shake,
+    module Rib.Route,
     MMark,
     Pandoc,
   )
@@ -10,4 +11,5 @@ where
 import Rib.App
 import Rib.Parser.MMark (MMark)
 import Rib.Parser.Pandoc (Pandoc)
+import Rib.Route
 import Rib.Shake